requirements.txt
Flask==2.3.3
pandas==2.1.1
openpyxl==3.1.2
sarvamai==0.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
numpy==1.24.3
